Change transfer amount JSON to string encoded num (#41)
This PR wraps the type of the transfer amount field in order to change its JSON representation to be a string literal containing a number (as opposed to a JSON number literal). 

The reason for this change is that [even though JSON-by-the-spec number literals are arbitrary-precision](https://github.com/kadena-io/chainweb-api/pull/30#issuecomment-1255044266), we've concluded that in practice, parsing JSON number literals with more precision than JS numbers can be unnecessarily hard and fragile using the JSON libraries of many popular languages.

Seems like this representation is commonplace in the blockchain world anyway, so it shouldn't be surprising for most of our users.